TaskModel moved to separate file
